Rename tracking function file and moved hook to the file.
